api/requirements.txt
SQLAlchemy==1.4.41
virtualenv==20.16.5
prospector==1.7.7
pre-commit==2.20.0
Flask==2.2.2
mysqlclient==2.1.1
requests==2.28.1
pytest==7.1.3
connexion==2.14.1
Werkzeug==2.2.2
flask-cors==3.0.10
mysqlclient==2.1.1
mypy==0.982
DateTime==4.7
python-dateutil==2.8.2
pytest-mock==3.10.0